Solucion problema camara, deleteo de usuario junto a sus posteos, creacion de coleccion de posteos y modificacion de estilos acordes en la screen comments

+++ src/components/Post.js
@@ -89,7 +89,11 @@ class Post extends Component {
     }
 
     deletePost() {
-        db.collection('posts').doc(this.props.id).delete()
+        if (confirm('Delete post') === true){
+            db.collection('posts').doc(this.props.id).delete()
+        } else {
+            false
+        }
     }
 
     setCommentsCount(num) {
